+++ include/usermanager.h
@@ -158,32 +158,32 @@ class CoreExport UserManager : public fakederef<UserManager>
 	/** Return a count of all global users, unknown and known connections
 	 * @return The number of users on the network, including local unregistered users
 	 */
-	unsigned int UserCount() const { return this->clientlist.size(); }
+	size_t UserCount() const { return this->clientlist.size(); }
 
 	/** Return a count of fully registered connections on the network
 	 * @return The number of registered users on the network
 	 */
-	unsigned int RegisteredUserCount() { return this->clientlist.size() - this->UnregisteredUserCount() - this->ULineCount(); }
+	size_t RegisteredUserCount() { return this->clientlist.size() - this->UnregisteredUserCount() - this->ULineCount(); }
 
 	/** Return a count of opered (umode +o) users on the network
 	 * @return The number of opers on the network
 	 */
-	unsigned int OperCount() const { return this->all_opers.size(); }
+	size_t OperCount() const { return this->all_opers.size(); }
 
 	/** Return a count of local unregistered (before NICK/USER) users
 	 * @return The number of local unregistered (unknown) connections
 	 */
-	unsigned int UnregisteredUserCount() const { return this->unregistered_count; }
+	size_t UnregisteredUserCount() const { return this->unregistered_count; }
 
 	/** Return a count of users on a u-lined servers.
 	 * @return The number of users on u-lined servers.
 	 */
-	unsigned int ULineCount() const { return this->all_ulines.size(); }
+	size_t ULineCount() const { return this->all_ulines.size(); }
 
 	/** Return a count of local registered users
 	 * @return The number of registered local users
 	 */
-	unsigned int LocalUserCount() const { return (this->local_users.size() - this->UnregisteredUserCount()); }
+	size_t LocalUserCount() const { return (this->local_users.size() - this->UnregisteredUserCount()); }
 
 	/** Get a hash map containing all users, keyed by their nickname
 	 * @return A hash map mapping nicknames to User pointers
