+++ src/mode.cpp
@@ -365,7 +365,7 @@ ModeAction ModeParser::TryMode(User* user, User* targetuser, Channel* chan, Mode
 	return MODEACTION_ALLOW;
 }
 
-void ModeParser::ModeParamsToChangeList(User* user, ModeType type, const std::vector<std::string>& parameters, Modes::ChangeList& changelist, unsigned int beginindex, unsigned int endindex)
+void ModeParser::ModeParamsToChangeList(User* user, ModeType type, const std::vector<std::string>& parameters, Modes::ChangeList& changelist, size_t beginindex, size_t endindex)
 {
 	if (endindex > parameters.size())
 		endindex = parameters.size();
@@ -373,7 +373,7 @@ void ModeParser::ModeParamsToChangeList(User* user, ModeType type, const std::ve
 	const std::string& mode_sequence = parameters[beginindex];
 
 	bool adding = true;
-	unsigned int param_at = beginindex+1;
+	size_t param_at = beginindex+1;
 
 	for (std::string::const_iterator letter = mode_sequence.begin(); letter != mode_sequence.end(); letter++)
 	{
