+++ src/modules/m_blockamsg.cpp
@@ -50,7 +50,7 @@ class BlockedMessage
 
 class ModuleBlockAmsg : public Module
 {
-	unsigned int ForgetDelay;
+	unsigned long ForgetDelay;
 	BlockAction action;
 	SimpleExtItem<BlockedMessage> blockamsg;
 
@@ -118,7 +118,12 @@ class ModuleBlockAmsg : public Module
 			// OR
 			// The number of target channels is equal to the number of channels the sender is on..a little suspicious.
 			// Check it's more than 1 too, or else users on one channel would have fun.
-			if ((m && (m->message == parameters[1]) && (!irc::equals(m->target, parameters[0])) && ForgetDelay && (m->sent >= ServerInstance->Time()-ForgetDelay)) || ((targets > 1) && (targets == user->chans.size())))
+			if ((m &&
+						(m->message == parameters[1]) &&
+						(!irc::equals(m->target, parameters[0])) &&
+						ForgetDelay &&
+						(m->sent >= ServerInstance->Time()-(long)ForgetDelay)) ||
+					((targets > 1) && (targets == user->chans.size())))
 			{
 				// Block it...
 				if (action == IBLOCK_KILLOPERS || action == IBLOCK_NOTICEOPERS)
