+++ src/modules/m_conn_join.cpp
@@ -80,7 +80,7 @@ class ModuleConnJoin : public Module
 	{
 		ConfigTag* tag = ServerInstance->Config->ConfValue("autojoin");
 		defchans = tag->getString("channel");
-		defdelay = tag->getDuration("delay", 0, 0, 60*15);
+		defdelay = static_cast<unsigned int>(tag->getDuration("delay", 0, 0, 60*15));
 	}
 
 	void Prioritize() CXX11_OVERRIDE
@@ -100,7 +100,7 @@ class ModuleConnJoin : public Module
 			return;
 
 		std::string chanlist = localuser->GetClass()->config->getString("autojoin");
-		unsigned int chandelay = localuser->GetClass()->config->getDuration("autojoindelay", 0, 0, 60*15);
+		unsigned int chandelay = static_cast<unsigned int>(localuser->GetClass()->config->getDuration("autojoindelay", 0, 0, 60*15));
 
 		if (chanlist.empty())
 		{
